+++ tests/testthat/test-spline_monotone.R
@@ -156,7 +156,7 @@ test_that("tunable", {
 
   rec <-
     recipe(~., data = iris) %>%
-      step_spline_monotone(all_predictors())
+    step_spline_monotone(all_predictors())
   rec_param <- tunable.step_spline_monotone(rec$steps[[1]])
   expect_equal(rec_param$name, c("deg_free", "degree"))
   expect_true(all(rec_param$source == "recipe"))
@@ -178,6 +178,18 @@ test_that("works when baked with 1 row", {
   )
 
   expect_identical(nrow(res), 1L)
+})
+
+test_that("errors with zero variance predictors (#1455)", {
+  mtcars$disp <- 1
+  mtcars$vs <- 1
+
+  expect_snapshot(
+    error = TRUE,
+    recipe(mpg ~ ., data = mtcars) %>%
+      step_spline_monotone(all_numeric_predictors()) %>%
+      prep()
+  )
 })
 
 # Infrastructure ---------------------------------------------------------------